[fix] Add monkeypatch for AES encryption
The upstream aes gem has an [open issue](https://github.com/chicks/aes/issues/17) regarding
the key size that is set internally. This results in an exception from
OpenSSL.
A PR to fix the issue is already available, but it has not yet been
merged. I decided to add a workaround via a monkey patch for the time
being. This way, we don't need to wait for the upstream merge to fix our
issue.

Once that ticket is closed, the monkeypatch can be removed.
closes #45